+++ FMSystems.WeatherForecast.Domain/FMSystems.WeatherForecast.Domain.csproj
@@ -7,4 +7,8 @@
   <ItemGroup>
     <SupportedPlatform Include="browser" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Service\" />
+  </ItemGroup>
 </Project>
