--- FMSystems.WeatherForecast.Infrastructure/Db/Repository/GenericDbRepository.cs
@@ -1,101 +1,0 @@
-﻿using FMSystems.WeatherForecast.Domain;
-using FMSystems.WeatherForecast.Domain.Entity;
-using FMSystems.WeatherForecast.Infrastructure.DBContext;
-using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-
-namespace FMSystems.WeatherForecast.Infrastructure.Db.Repository
-{
-    public class GenericDbRepository<TEntity> : IDisposable, IGenericDbRepository<TEntity> where TEntity : BaseEntity
-    {
-        internal WeatherForecastDbContext context;
-        internal DbSet<TEntity> dbSet;
-        private bool disposed = false;
-
-        public GenericDbRepository(WeatherForecastDbContext context)
-        {
-            this.context = context;
-            this.dbSet = context.Set<TEntity>();
-            this.context.Database.EnsureCreated();
-        }
-
-        public async virtual Task<IEnumerable<TEntity>> GetAsync(
-            Expression<Func<TEntity, bool>> filter = null,
-            Func<IQueryable<TEntity>, IOrderedQueryable<TEntity>> orderBy = null)
-        {
-            IQueryable<TEntity> query = dbSet;
-
-            if (filter != null)
-            {
-                query = query.Where(filter);
-            }
-
-            if (orderBy != null)
-            {
-                return await orderBy(query).ToListAsync();
-            }
-            else
-            {
-                return await query.ToListAsync();
-            }
-        }
-
-        public async virtual Task<TEntity> GetByIDAsync(object id)
-        {
-            return await dbSet.FindAsync(id);
-        }
-
-        public async virtual Task InsertAsync(TEntity entity)
-        {
-            await dbSet.AddAsync(entity);
-        }
-
-        public async virtual Task DeleteAsync(object id)
-        {
-            TEntity entityToDelete = await dbSet.FindAsync(id);
-            Delete(entityToDelete);
-        }
-
-        public virtual void Delete(TEntity entityToDelete)
-        {
-            if (context.Entry(entityToDelete).State == EntityState.Detached)
-            {
-                dbSet.Attach(entityToDelete);
-            }
-            dbSet.Remove(entityToDelete);
-        }
-
-        public virtual void Update(TEntity entityToUpdate)
-        {
-            dbSet.Attach(entityToUpdate);
-            context.Entry(entityToUpdate).State = EntityState.Modified;
-        }
-
-        public async virtual Task SaveAsync()
-        {
-            await context.SaveChangesAsync();
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!this.disposed)
-            {
-                if (disposing)
-                {
-                    context.Dispose();
-                }
-            }
-            this.disposed = true;
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-    }
-}
